--- src/main/java/com/wearedevs/config/security/provider/CustomProvider.java
@@ -1,5 +1,0 @@
-package com.wearedevs.config.security.provider;
-
-public interface CustomProvider {
-
-}
